requirements.txt
Pillow==8.3.0
pytesseract==0.3.8
pyttsx3==2.90
ttkbootstrap==0.5.1
